if hmac path doesn't not exist, don't try to go further

+++ lib/index.py
@@ -63,6 +63,9 @@ class manage:
         if filepath is None:
             return False
 
+        if not os.path.exists(filepath):
+            return False
+
         tlocalfile = tmpname.get(filepath)
 
         f = open(tlocalfile[1], "w")
